% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_scores.R
\name{box_scores}
\alias{box_scores}
\title{A function to find box scores}
\usage{
box_scores(data = NULL, row = 1)
}
\arguments{
\item{box_scores, }{no defaults}
}
\description{
A function to find box scores
}
\examples{
\dontrun{box_scores()}
box_scores()
}
\keyword{box,}
\keyword{box_scores,}
\keyword{scores}
