% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixtures.R
\name{fixtures}
\alias{fixtures}
\title{A function to find fixture results}
\usage{
fixtures(start_season = 1996, end_season = 2021)
}
\arguments{
\item{fixtures, }{defaults are start=1996 and end=2021}
}
\description{
A function to find fixture results
}
\examples{
library(rMLS)
fixtures(start_season=2008,end_season=2020)
}
\keyword{fixtures,}
\keyword{results,}
\keyword{schedule}
