% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fte_rankings.R
\name{fte_rankings}
\alias{fte_rankings}
\title{A function to pull fivethirtyeight's global club rankings}
\usage{
fte_rankings()
}
\arguments{
\item{fte_rankings, }{defaults to 2017,2020}
}
\description{
This function allows me to test setup
}
\examples{
\dontrun{fte_rankings()}
fte_rankings()
}
\keyword{fivethirtyeight,}
\keyword{fte,}
\keyword{rankings}
