% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_stats.R
\name{player_stats}
\alias{player_stats}
\title{A function to find individual player statistics}
\usage{
player_stats(id = NULL)
}
\description{
This function allows me to test setup
}
\examples{
\dontrun{player_stats()}
player_stats()
}
\keyword{player,}
\keyword{player_stats}
\keyword{statistics,}
\keyword{stats,}
