% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playerid_lookup.R
\name{playerid_lookup}
\alias{playerid_lookup}
\title{A function to find the identification number for any player}
\usage{
playerid_lookup(id = NULL, name = NULL)
}
\arguments{
\item{playerid_lookup, }{no default, pulls all players}
}
\description{
This function allows me to test setup
}
\examples{
\dontrun{playerid_lookup()}
playerid_lookup()
}
\keyword{id,}
\keyword{lookup}
\keyword{player,}
\keyword{playerid,}
