% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmls.R
\name{rmls}
\alias{rmls}
\title{rmls}
\usage{
rmls(df = .Last.value, Y.var = "", Time.var = "",
  Current.time = "", launch.browser = False)
}
\arguments{
\item{df}{A dataframe input}

\item{Y.var}{Specify value dependent variable}

\item{Time.var}{Specify time variables such as Year, Quarter, and Month}

\item{Current.time}{Specify current time such as 2018}
}
\description{
rmls
}
