% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosters.R
\name{rosters}
\alias{rosters}
\title{A function to find fixture results}
\usage{
rosters(start_season = 1996, end_season = 2020, teamid = NULL)
}
\arguments{
\item{rosters, }{defaults are start=1996 and end=2021}
}
\description{
A function to find fixture results
}
\examples{
\dontrun{rosters()}
rosters()
}
\keyword{rosters,roster,team}
