% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_stats.R
\name{team_stats}
\alias{team_stats}
\title{A function to find team statistics}
\usage{
team_stats(start_season = 1996, end_season = 2020, team_name = NULL)
}
\arguments{
\item{team_stats, }{no defaults}
}
\description{
A function to find team statistics
}
\examples{
\dontrun{team_stats()}
team_stats()
}
\keyword{statistics,}
\keyword{stats,}
\keyword{team,}
\keyword{team_stats}
