% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_stats.R
\name{player_stats}
\alias{player_stats}
\title{A function to find individual player statistics}
\usage{
player_stats(id = NULL)
}
\arguments{
\item{player_stats, }{ID = Football Reference ID}
}
\description{
This function pulls individual player statistics using the Football-Reference player ID as an argument. An ID must be specified as there are no default arguments. Player ID values can be found using the playerid_lookup() function, where you can supply a name to generate a tibble of possible players.
}
\keyword{player,}
\keyword{player_stats}
\keyword{statistics,}
\keyword{stats,}
